Move data loading function to conftest.py and remove datasets module
